refactor: WI-0295 core decomposition phase1
