+++ src/app/admin/page-helpers.ts
@@ -1,0 +1,74 @@
+export function isTruthyFlag(value: string | undefined) {
+  const normalized = (value ?? "").trim().toLowerCase();
+  return normalized === "1" || normalized === "true" || normalized === "yes" || normalized === "on";
+}
+
+export function toLocalInputValue(value: Date) {
+  const adjusted = new Date(value.getTime() - value.getTimezoneOffset() * 60_000);
+  return adjusted.toISOString().slice(0, 16);
+}
+
+export function firstDayOfMonthLocal() {
+  const now = new Date();
+  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth(), 1, 0, 0, 0));
+}
+
+export function lastDayOfMonthLocal() {
+  const now = new Date();
+  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth() + 1, 0, 23, 59, 0));
+}
+
+export function toIso(value: string) {
+  return new Date(value).toISOString();
+}
+
+export function formatDateTime(value: string | null) {
+  if (!value) {
+    return "-";
+  }
+  const parsed = new Date(value);
+  if (Number.isNaN(parsed.getTime())) {
+    return value;
+  }
+  return parsed.toLocaleString("ko-KR");
+}
+
+export function formatDays(value: number) {
+  return Number.isInteger(value) ? String(value) : value.toFixed(2);
+}
+
+export function buildQuery(params: Record<string, string | undefined>) {
+  const search = new URLSearchParams();
+  for (const [key, value] of Object.entries(params)) {
+    if (value === undefined || value.trim() === "") {
+      continue;
+    }
+    search.set(key, value);
+  }
+  const query = search.toString();
+  return query.length > 0 ? `?${query}` : "";
+}
+
+export function minutesToHours(minutes: number) {
+  const hours = minutes / 60;
+  return `${hours.toFixed(1)}h`;
+}
+
+export function toTimestamp(value: string | null) {
+  if (!value) {
+    return 0;
+  }
+  const parsed = new Date(value);
+  if (Number.isNaN(parsed.getTime())) {
+    return 0;
+  }
+  return parsed.getTime();
+}
+
+export function toWaitHours(value: string | null, referenceMs: number) {
+  const timestamp = toTimestamp(value);
+  if (timestamp <= 0) {
+    return 0;
+  }
+  return Math.max(0, (referenceMs - timestamp) / 3_600_000);
+}
