+++ src/app/admin/page-types.ts
@@ -1,0 +1,115 @@
+export type ApiLog = {
+  id: number;
+  label: string;
+  status: number;
+  ok: boolean;
+  durationMs: number;
+  at: string;
+  body: unknown;
+};
+
+export type EmployeeSummary = {
+  id: string;
+  organizationId: string | null;
+  name: string | null;
+  email: string | null;
+  active: boolean;
+};
+
+export type OrganizationSummary = {
+  id: string;
+  name: string;
+};
+
+export type AttendanceRecordDto = {
+  id: string;
+  employeeId: string;
+  checkInAt: string;
+  checkOutAt: string | null;
+  breakMinutes: number;
+  isHoliday: boolean;
+  notes: string | null;
+  state: "PENDING" | "APPROVED" | "REJECTED";
+};
+
+export type WorkScheduleDto = {
+  id: string;
+  employeeId: string;
+  startAt: string;
+  endAt: string;
+  breakMinutes: number;
+  isHoliday: boolean;
+  notes: string | null;
+};
+
+export type InviteRole = "admin" | "manager" | "employee" | "payroll_operator";
+export type InviteDeliveryMode = "link" | "email";
+
+export type InviteResultDto = {
+  userId: string;
+  email: string;
+  role: InviteRole;
+  organizationId: string;
+  actorId: string | null;
+  redirectTo: string;
+  deliveryMode: InviteDeliveryMode;
+  actionLink: string | null;
+};
+
+export type LeaveRequestDto = {
+  id: string;
+  employeeId: string;
+  leaveType: "ANNUAL" | "SICK" | "UNPAID";
+  startDate: string;
+  endDate: string;
+  unit: "FULL_DAY" | "HALF_DAY" | "HOUR";
+  hours: number | null;
+  days: number;
+  reason: string | null;
+  decisionReason: string | null;
+  state: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
+};
+
+export type PayrollRunDto = {
+  id: string;
+  organizationId: string | null;
+  employeeId: string | null;
+  periodStart: string;
+  periodEnd: string;
+  state: "PREVIEWED" | "CONFIRMED";
+  grossPayKrw: number;
+  totalDeductionsKrw: number | null;
+  netPayKrw: number | null;
+  sourceRecordCount: number;
+  confirmedAt: string | null;
+  confirmedBy: string | null;
+};
+
+export type AttendanceAggregateDto = {
+  employeeId: string;
+  periodStart: string;
+  periodEnd: string;
+  counts: {
+    total: number;
+    pending: number;
+    approved: number;
+    rejected: number;
+    payable: number;
+  };
+  totals: {
+    regular: number;
+    overtime: number;
+    night: number;
+    holiday: number;
+  };
+};
+
+export type LeaveBalanceDto = {
+  employeeId: string;
+  grantedDays: number;
+  usedDays: number;
+  remainingDays: number;
+  carryOverDays: number;
+  lastAccrualYear: number | null;
+  updatedAt: string;
+};
